+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## v26.1.2.22
-* Improved itemslot/fluidslot drawing function overridable
-* Fixed JEI recipe slot size
+## v26.1.2.23
+* Fixed z-index draw
+* Fixed GraphView keydown event doesn’t use
+* Improved IDataConsumer + IObserbale apis. + Added xei shift pause scroll